refactor: move TextMatchOptiond'
