+++ src/queries/placeholderText.ts
@@ -12,7 +12,7 @@ import type {
   QueryAllByQuery,
   QueryByQuery,
 } from './makeQueries';
-import type { TextMatchOptions } from './text';
+import type { TextMatchOptions } from './options';
 
 const getTextInputNodeByPlaceholderText = (
   node: ReactTestInstance,
